requirements.txt
# AF-Training Requirements
# For compatibility with pip-based workflows
# Prefer using: uv sync

# Core ML
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision  
opencv-python>=4.8.0
pillow>=10.0.0

# ONNX & Export
onnx>=1.14.0
onnxsim>=0.4.33

# Data handling
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress & Logging
tqdm>=4.65.0

# Note: TensorRT requires manual installation from NVIDIA
# For WSL2/Ubuntu: https://docs.nvidia.com/deeplearning/tensorrt/install-guide/index.html
# Or use the Docker container which has it pre-installed
